cloud_functions/requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-storage>=2.13.0
functions-framework>=3.0.0
